// darkhorse-java/toolkit/packages/rich-text-editor/src/types.ts
export type RteBuiltinCommand =
  | 'Undo'
  | 'Redo'
  | 'Bold'
  | 'Italic'
  | 'Underline'
  | 'StrikeThrough'
  | 'InlineCode'
  | 'Superscript'
  | 'Subscript'
  | 'FontColor'
  | 'BackgroundColor'
  | 'FontName'
  | 'FontSize'
  | 'Formats'
  | 'AlignLeft'
  | 'AlignCenter'
  | 'AlignRight'
  | 'AlignJustify'
  | 'OrderedList'
  | 'UnorderedList'
  | 'Checklist'
  | 'Outdent'
  | 'Indent'
  | 'Blockquote'
  | 'HorizontalLine'
  | 'CreateLink'
  | 'RemoveLink'
  | 'InsertImage'
  | 'InsertTable'
  | 'InsertVideo'
  | 'InsertAudio'
  | 'EmojiPicker'
  | 'FormatPainter'
  | 'ClearFormat'
  | 'Print'
  | 'FullScreen'
  | 'SourceCode'
  | 'ImportWord'
  | 'ExportWord'
  | 'ExportPdf';